requirements.txt
attrs==18.2.0
Automat==0.7.0
certifi==2018.10.15
chardet==3.0.4
constantly==15.1.0
cycler==0.10.0
fredapi==0.4.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
inflection==0.3.1
kiwisolver==1.0.1
llvmlite==0.25.0
matplotlib==3.0.1
more-itertools==4.3.0
numba==0.40.1
numpy==1.15.4
pandas==0.23.4
PyHamcrest==1.9.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
Quandl==3.4.4
requests==2.20.0
scipy==1.1.0
six==1.11.0
Twisted==18.9.0
urllib3==1.24.1
zope.interface==4.6.0
